Refactor business logic out of the appdistribution CLI so that it can be used by an MCP tool.
